Added support for 'copy' stream
'copy' means that FFMPEG will do as little as possible to copy the
stream input to desired output. It means that one can't apply filters or
store the output in a format that can't be directly converted from the
input, but it removes all overhead.

This commit also fixes a bug in `size` which resulted in the input
format to always being the smallest possible.